added more test-cases
